// backend_microservices/account-status-service/account_status_service/src/enums/status-change-reason.enum.ts
export enum StatusChangeReason {
  FRAUD_INVESTIGATION = 'fraud_investigation',
  CUSTOMER_REQUEST = 'customer_request',
  DELINQUENT_ACCOUNT = 'delinquent_account',
  ACCOUNT_UPGRADE = 'account_upgrade',
  REGULATORY_COMPLIANCE = 'regulatory_compliance',
  SYSTEM_MAINTENANCE = 'system_maintenance',
}